+++ OniAccess/Handlers/Screens/ConsumablesScreenHandler.cs
@@ -129,7 +129,7 @@ namespace OniAccess.Handlers.Screens {
 					_rows.Add(new RowEntry { Kind = TableRowKind.WorldDivider, WorldId = worldId });
 
 				foreach (var minion in minions) {
-					_rows.Add(new RowEntry { Kind = TableRowKind.Minion, Identity = minion });
+					_rows.Add(new RowEntry { Kind = TableRowKind.Minion, Identity = minion, WorldId = worldId });
 				}
 			}
 
@@ -138,7 +138,7 @@ namespace OniAccess.Handlers.Screens {
 				if (showDividers)
 					_rows.Add(new RowEntry { Kind = TableRowKind.WorldDivider, WorldId = StoredMinionWorldId });
 				foreach (var smi in stored) {
-					_rows.Add(new RowEntry { Kind = TableRowKind.StoredMinion, Identity = smi });
+					_rows.Add(new RowEntry { Kind = TableRowKind.StoredMinion, Identity = smi, WorldId = StoredMinionWorldId });
 				}
 			}